clean codecheck

+++ interfaces/kits/js/src/mod_fileio/class_dir/dir_n_exporter.cpp
@@ -38,7 +38,7 @@ using namespace std;
 static DirEntity *GetDirEntity(napi_env env, napi_callback_info info)
 {
     NFuncArg funcArg(env, info);
-    if (!funcArg.InitArgs(NARG_CNT::ZERO)) {
+    if (!funcArg.InitArgs(static_cast<size_t>(NARG_CNT::ZERO))) {
         UniError(EINVAL).ThrowErr(env, "Number of arguments unmatched");
         return nullptr;
     }
@@ -66,7 +66,7 @@ napi_value DirNExporter::CloseSync(napi_env env, napi_callback_info info)
 napi_value DirNExporter::Close(napi_env env, napi_callback_info info)
 {
     NFuncArg funcArg(env, info);
-    if (!funcArg.InitArgs(NARG_CNT::ZERO, NARG_CNT::ONE)) {
+    if (!funcArg.InitArgs(static_cast<size_t>(NARG_CNT::ZERO), static_cast<size_t>(NARG_CNT::ONE))) {
         UniError(EINVAL).ThrowErr(env, "Number of arguments unmatched");
         return nullptr;
     }
@@ -101,10 +101,10 @@ napi_value DirNExporter::Close(napi_env env, napi_callback_info info)
 
     NVal thisVar(env, funcArg.GetThisVar());
     string procedureName = "fileioDirClose";
-    if (funcArg.GetArgc() == NARG_CNT::ZERO) {
+    if (funcArg.GetArgc() == static_cast<size_t>(NARG_CNT::ZERO)) {
         return NAsyncWorkPromise(env, thisVar).Schedule(procedureName, cbExec, cbCompl).val_;
     } else {
-        NVal cb(env, funcArg[NARG_POS::FIRST]);
+        NVal cb(env, funcArg[static_cast<size_t>(NARG_POS::FIRST)]);
         return NAsyncWorkCallback(env, thisVar, cb).Schedule(procedureName, cbExec, cbCompl).val_;
     }
 }
@@ -142,7 +142,7 @@ static NVal DoReadCompile(napi_env env, UniError err, shared_ptr<DirReadArgs> ar
 napi_value DirNExporter::Read(napi_env env, napi_callback_info info)
 {
     NFuncArg funcArg(env, info);
-    if (!funcArg.InitArgs(NARG_CNT::ZERO, NARG_CNT::ONE)) {
+    if (!funcArg.InitArgs(static_cast<size_t>(NARG_CNT::ZERO), static_cast<size_t>(NARG_CNT::ONE))) {
         UniError(EINVAL).ThrowErr(env, "Number of arguments unmatched");
         return nullptr;
     }
@@ -187,10 +187,10 @@ napi_value DirNExporter::Read(napi_env env, napi_callback_info info)
     };
     NVal thisVar(env, funcArg.GetThisVar());
 
-    if (funcArg.GetArgc() == NARG_CNT::ZERO) {
+    if (funcArg.GetArgc() == static_cast<size_t>(NARG_CNT::ZERO)) {
         return NAsyncWorkPromise(env, thisVar).Schedule("fileioDirRead", cbExec, cbCompl).val_;
     } else {
-        NVal cb(env, funcArg[NARG_POS::FIRST]);
+        NVal cb(env, funcArg[static_cast<size_t>(NARG_POS::FIRST)]);
         return NAsyncWorkCallback(env, thisVar, cb).Schedule("fileioDirRead", cbExec, cbCompl).val_;
     }
 }
@@ -198,7 +198,7 @@ napi_value DirNExporter::Read(napi_env env, napi_callback_info info)
 napi_value DirNExporter::ReadSync(napi_env env, napi_callback_info info)
 {
     NFuncArg funcArg(env, info);
-    if (!funcArg.InitArgs(NARG_CNT::ZERO)) {
+    if (!funcArg.InitArgs(static_cast<size_t>(NARG_CNT::ZERO))) {
         UniError(EINVAL).ThrowErr(env, "Number of arguments unmatched");
         return nullptr;
     }
@@ -247,7 +247,7 @@ napi_value DirNExporter::ReadSync(napi_env env, napi_callback_info info)
 napi_value DirNExporter::Constructor(napi_env env, napi_callback_info info)
 {
     NFuncArg funcArg(env, info);
-    if (!funcArg.InitArgs(NARG_CNT::ZERO)) {
+    if (!funcArg.InitArgs(static_cast<size_t>(NARG_CNT::ZERO))) {
         UniError(EINVAL).ThrowErr(env, "Number of arguments unmatched");
         return nullptr;
     }
