+++ interfaces/kits/js/src/mod_fileio/class_stream/flush.cpp
@@ -39,7 +39,7 @@ struct StreamEntity {
 napi_value Flush::Sync(napi_env env, napi_callback_info info)
 {
     NFuncArg funcArg(env, info);
-    if (!funcArg.InitArgs(NARG_CNT::ZERO)) {
+    if (!funcArg.InitArgs(static_cast<size_t>(NARG_CNT::ZERO))) {
         UniError(EINVAL).ThrowErr(env, "Number of arguments unmatched");
         return nullptr;
     }
@@ -61,7 +61,7 @@ napi_value Flush::Sync(napi_env env, napi_callback_info info)
 napi_value Flush::Async(napi_env env, napi_callback_info info)
 {
     NFuncArg funcArg(env, info);
-    if (!funcArg.InitArgs(NARG_CNT::ZERO, NARG_CNT::ONE)) {
+    if (!funcArg.InitArgs(static_cast<size_t>(NARG_CNT::ZERO), static_cast<size_t>(NARG_CNT::ONE))) {
         UniError(EINVAL).ThrowErr(env, "Number of arguments unmatched");
         return nullptr;
     }
@@ -89,10 +89,10 @@ napi_value Flush::Async(napi_env env, napi_callback_info info)
 
     NVal thisVar(env, funcArg.GetThisVar());
     string procedureName = "fileIOFlush";
-    if (funcArg.GetArgc() == NARG_CNT::ZERO) {
+    if (funcArg.GetArgc() == static_cast<size_t>(NARG_CNT::ZERO)) {
         return NAsyncWorkPromise(env, thisVar).Schedule(procedureName, cbExec, cbCompl).val_;
     } else {
-        NVal cb(env, funcArg[NARG_POS::FIRST]);
+        NVal cb(env, funcArg[static_cast<size_t>(NARG_POS::FIRST)]);
         return NAsyncWorkCallback(env, thisVar, cb).Schedule(procedureName, cbExec, cbCompl).val_;
     }
 }
