+++ interfaces/kits/js/src/mod_fileio/properties/fdopen_stream.cpp
@@ -53,14 +53,14 @@ static tuple<bool, int, string> GetFdopenStreamArgs(napi_env env, const NFuncArg
 {
     bool succ = false;
     int fd;
-    tie(succ, fd) = NVal(env, funcArg[NARG_POS::FIRST]).ToInt32();
+    tie(succ, fd) = NVal(env, funcArg[static_cast<size_t>(NARG_POS::FIRST)]).ToInt32();
     if (!succ) {
         UniError(EINVAL).ThrowErr(env, "Arg fd is required to be type integer");
         return { false, -1, "" };
     }
 
     unique_ptr<char[]> mode;
-    tie(succ, mode, ignore) = NVal(env, funcArg[NARG_POS::SECOND]).ToUTF8String();
+    tie(succ, mode, ignore) = NVal(env, funcArg[static_cast<size_t>(NARG_POS::SECOND)]).ToUTF8String();
     if (!succ) {
         UniError(EINVAL).ThrowErr(env, "Arg mode is required to be type string");
         return { false, -1, "" };
@@ -72,7 +72,7 @@ napi_value FdopenStream::Sync(napi_env env, napi_callback_info info)
 {
     NFuncArg funcArg(env, info);
 
-    if (!funcArg.InitArgs(NARG_CNT::TWO)) {
+    if (!funcArg.InitArgs(static_cast<size_t>(NARG_CNT::TWO))) {
         UniError(EINVAL).ThrowErr(env, "Number of arguments unmatched");
         return nullptr;
     }
@@ -102,7 +102,7 @@ struct AsyncFdopenStreamArg {
 napi_value FdopenStream::Async(napi_env env, napi_callback_info info)
 {
     NFuncArg funcArg(env, info);
-    if (!funcArg.InitArgs(NARG_CNT::TWO, NARG_CNT::THREE)) {
+    if (!funcArg.InitArgs(static_cast<size_t>(NARG_CNT::TWO), static_cast<size_t>(NARG_CNT::THREE))) {
         UniError(EINVAL).ThrowErr(env, "Number of arguments unmatched");
         return nullptr;
     }
@@ -134,10 +134,10 @@ napi_value FdopenStream::Async(napi_env env, napi_callback_info info)
 
     string procedureName = "FileIOFdopenStream";
     NVal thisVar(env, funcArg.GetThisVar());
-    if (funcArg.GetArgc() == NARG_CNT::TWO) {
+    if (funcArg.GetArgc() == static_cast<size_t>(NARG_CNT::TWO)) {
         return NAsyncWorkPromise(env, thisVar).Schedule(procedureName, cbExec, cbCompl).val_;
     } else {
-        NVal cb(env, funcArg[NARG_POS::THIRD]);
+        NVal cb(env, funcArg[static_cast<size_t>(NARG_POS::THIRD)]);
         return NAsyncWorkCallback(env, thisVar, cb).Schedule(procedureName, cbExec, cbCompl).val_;
     }
 }
